requirements.txt
asgiref==3.2.10
Django==3.1
django-crispy-forms==1.9.2
django-extensions==3.1.1
django-tinymce==3.2.0
Pillow==8.0.1
pytz==2020.1
sqlparse==0.3.1